+++ core/log.c
@@ -572,10 +572,12 @@ static i32 try_init_new_output(enum s_log_level level,
 
 static void destroy_old_output(struct output *o)
 {
-    if (o->type == S_LOG_OUTPUT_FILEPATH || o->type == S_LOG_OUTPUT_FILE) {
-        fflush(o->fp);
-        fclose(o->fp);
-    }
+    if (o->type == S_LOG_OUTPUT_FILEPATH  || o->type == S_LOG_OUTPUT_FILE)
+        (void) fflush(o->fp);
+
+    if (o->type == S_LOG_OUTPUT_FILEPATH)
+        (void) fclose(o->fp);
+
     memset(o, 0, sizeof(struct output));
 }