+++ game/init.c
@@ -7,32 +7,18 @@
 #include <platform/window.h>
 #include <platform/keyboard.h>
 #include <platform/platform.h>
-#include <string.h>
 #define CGD_GAME_CONFIG__
 #include "config.h"
 #undef CGD_GAME_CONFIG__
 
 #define MODULE_NAME "init"
 
-static i32 setup_log(void);
-
 i32 do_platform_init(i32 argc, const char *const *argv,
     struct platform_ctx *ctx)
 {
     (void) argc;
-    s_log_info("Starting platform init...");
-
-    /* Set up logging */
-#ifndef CGD_BUILDTYPE_RELEASE
-    if (!strcmp(argv[0], "debug"))
-        s_configure_log_level(S_LOG_TRACE);
-#else
-    s_configure_log_level(S_LOG_INFO);
     (void) argv;
-#endif /* CGD_BUILDTYPE_RELEASE */
-    if (setup_log())
-        goto err;
-    s_log_verbose("Log setup OK");
+    s_log_info("Starting platform init...");
 
     s_log_verbose("Creating the window...");
     ctx->win = p_window_open(WINDOW_TITLE, &WINDOW_RECT, WINDOW_FLAGS);
@@ -86,50 +72,10 @@ void do_platform_cleanup(struct platform_ctx *ctx)
     if (ctx->mouse != NULL) p_mouse_destroy(&ctx->mouse);
     if (ctx->keyboard != NULL) p_keyboard_destroy(&ctx->keyboard);
     if (ctx->win != NULL) p_window_close(&ctx->win);
-
-    s_log_cleanup_all();
 }
 
 void do_gui_cleanup(struct gui_ctx *gui)
 {
     if (gui->mmgr != NULL) menu_mgr_destroy(&gui->mmgr);
     if (gui->r != NULL) r_ctx_destroy(&gui->r);
-}
-
-static i32 setup_log(void)
-{
-#if (PLATFORM == PLATFORM_WINDOWS)
-#include <errno.h>
-
-    const struct s_log_output_cfg log_cfg = {
-        .type = S_LOG_OUTPUT_FILEPATH,
-        .out = { .filepath = LOG_FILEPATH },
-        .flag_copy = true,
-        .flag_append = true,
-    };
-    if (s_configure_log_outputs(S_LOG_ALL_MASKS, &log_cfg)) {
-        s_log_error("Failed to set the log configs to file \"%s\"",
-            LOG_FILEPATH);
-        return 1;
-    }
-
-#else
-    struct s_log_output_cfg log_cfg = {
-        .type = S_LOG_OUTPUT_FILE,
-        .out = { .file = stdout },
-        .flag_copy = true,
-    };
-    if (s_configure_log_outputs(S_LOG_STDOUT_MASKS, &log_cfg)) {
-        s_log_error("Failed to set out log configs");
-        return 1;
-    }
-
-    log_cfg.out.file = stderr;
-    if (s_configure_log_outputs(S_LOG_STDERR_MASKS, &log_cfg)) {
-        s_log_error("Failed to set error log configs");
-        return 1;
-    }
-#endif /* PLATFORM */
-
-    return 0;
 }